#4 [FEAT] 자동로그인 에러 로직 추가
